% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSetPerm.R
\name{geneSetPerm}
\alias{geneSetPerm}
\title{Test gene set enrichment by permutating gene labels of statistics}
\usage{
geneSetPerm(stats, indList, Nsim = 9999)
}
\arguments{
\item{stats}{Statistics}

\item{indList}{a list of integers, indicating indices of genes of gene sets (index starts from 1, following R's convention)}

\item{Nsim}{number of simulations}
}
\value{
A data frame containg mean statistic, gene set size, and p-values
}
\description{
TODO
}
\examples{
set.seed(1887)
stats <- rnorm(1000)
gsList <- list(gs1=c(3,4,5), gs2=c(7,8,9))
geneSetPerm(stats, gsList, Nsim=9999)
gsList2 <- list(gs1=c(3,4,5), gs2=c(7,8,9), gs3=integer())
geneSetPerm(stats, gsList2, Nsim=9999)
gsList3 <- sample(1:1000, 200)
geneSetPerm(stats, gsList3, Nsim=9999)
}
\seealso{
\code{geneSetTest}, a R implementation in the limma package
}

